Update docker setups to fit the most recent additions from Rehr and my migrid code updates.
Simplified the user creation now that openid alias is supported directly in the migrid code even for gdp mode.

Synced new vars and to other env files and adjust accordingly.
Bump to recent svn and git versions with required changes.
Enable twofactor support in all containers - without enforcement.

Revert default env to python2 and svn or git edge, as experimental is intended
for for python3 and a few components have diverged to be incompatible with
python2. This divergence may increase unless we find a good way to centrally
select the python version used in the shebang env python lines in general and
from sshd subsys callouts:

Removed explicit GDP option to docker-entry script and use a dynamic lookup
instead for the only remaining part about adding IO service logins for the test
account.