Replace the problematic nested volume mounts with bind mounts and add a shared volume init container. This is as proposed in relation to PR43 and the fallout it caused regarding build errors and races in launch from pristine clone e.g. when running in the existing CI.
Introduces the shared volume pre-init container to eliminate start-up races.

Uses env variables to allow flexible vgrid_files_readonly and mig_system_run
locations, in order to hopefully address the needs originally motivating PR43.

Please note that this is just a properly branched repeat of the original work
in my own fork of the repo but updated to use the current version of the file.