Initial support for hotfix integration as outlined in milestone 6.
Adds a new `/hotfixes` folder in containers and `${PERSISTENT_ROOT}/hotfixes`
with `scripts` and `patches` in it on the host.
Adds a simple script to help apply hotfixes from a folder and installs it in
`/app/apply-hotfixes.sh` in all containers.
All containers mount `${PERSISTENT_ROOT}/hotfixes` on `/hotfixes` and
`docker-entry.sh` is extended to first run `/app/apply-hotfixes.sh`.
It traverses `/hotfixes` and sequentially applies any patches found in
`/hotfix/patches` and runs any scripts found in `/hotfix/scripts` before
proceeding with the usual container execution to effectively hotfix everything
inside the container before running.